refactor: move screens into own folders
