+++ source/screens/bookingScreens/BookingFormScreen/index.ts
@@ -1,0 +1,3 @@
+import BookingFormScreen from "./BookingFormScreen";
+
+export default BookingFormScreen;
