+++ source/screens/bookingScreens/BookingSummaryScreen/index.ts
@@ -1,0 +1,3 @@
+import BookingSummary from "./BookingSummary";
+
+export default BookingSummary;
