+++ source/screens/bookingScreens/CalendarScreen/index.ts
@@ -1,0 +1,3 @@
+import CalendarScreen from "./CalendarScreen";
+
+export default CalendarScreen;
